WIP: There is still a problem with dependecies
I think that the dependencies are downloaded correctly. The problem is that they are put in weird places and the `go build` command cannot find them in their `GO_ROOT` or something.

I believe that once we figure out how to tell Go where to look for those packages the problem will be solved.